Add armi.reactor.grids.locations
Provides abstract `LocationBase` as the interface defining base class.
Concrete subclasses `IndexLocation`, `MultiIndexLocation`, and
`CoordinateLocation` are also defined.

There is little to no implementation differences in this change. Only
that `LocationBase` is a proper abstract base class, so subclasses that
don't override the necessary abstract methods will need to change. Some
other key changes

- `LocationBase.__eq__` returns `NotImplemented` if the RHS argument
  is not a tuple nor a `LocationBase`

The four main classes defined are imported into `armi.reactor.grids`
to maintain back compatibility.
